requirements.txt
Django==4.1.4
django-mptt==0.14.0
djangorestframework==3.14.0
environ==1.0
numpy==1.23.5
openai==0.25.0
Pillow==9.3.0

django-filter~=22.1
